+++ GLFW_Test/src/Engine/Engine.h
@@ -32,7 +32,7 @@ namespace EngineCore {
 	private:
 		int m_frameCount = 0;
 		int m_framesPerSecond = 0;// m_framesPerSecond = m_frameCount - m_lastFrameCount;
-		float m_fpsCounter = 0;// checks if a second has passed
+		double m_fpsCounter = 0;// checks if a second has passed
 		int m_lastFrameCount = 0;// is the frame count from the last sec 
 
 		GLFWwindow* m_window = nullptr;
